refactor(atreus/app/comment/data): modify data layer data connection startup and cleanup logic
Modify Redis connections to be accessed as structures for easier use. According to the analysis of
the goroutine access data, each goroutine in the structure to access the attributes, are different,
there is no resource competition, so eliminate the lock.

BREAKING CHANGE: Improve concurrency efficiency by removing the performance limitations of locks